Track changes in predicates
Extend the Pred.t record to include a source path and a flag
indicating whether it is internally-defined, like in Proc.t.

Extend the heuristics in changeTracker.ml to also take into account
changed predicates. If a predicate changes, then all the predicates
that are transitively using it as well as all procedures that use
these predicates (either in their specs or within proof tactics) must
be marked as changed.